refactor: 페이지 디렉토리 변경(#72)
