+++ src/router/Router.tsx
@@ -7,9 +7,9 @@ import { FooterLayout } from "@/layouts/FooterLayout";
 import NewPostPage from "@/pages/photoFeed/NewPostPage";
 import { MyPage } from "@/pages/mypage/MyPage";
 import { EditInfoPage } from "@/pages/mypage/edit-info/EditInfoPage";
-import { LikedPostPage } from "@/pages/mypage/LikedPostPage";
-import { LikedPhotoLabPage } from "@/pages/mypage/LikedPhotoLabPage";
-import { MyPostPage } from "@/pages/mypage/MyPostPage";
+import { LikedPostPage } from "@/pages/mypage/tab/LikedPostPage";
+import { LikedPhotoLabPage } from "@/pages/mypage/tab/LikedPhotoLabPage";
+import { MyPostPage } from "@/pages/mypage/tab/MyPostPage";
 import FindPhotoLabPage from "@/pages/photoFeed/FindPhotoLabPage";
 import ReviewPhotoLabPage from "@/pages/photoFeed/ReviewPhotoLabPage";
 import MyPageLayout from "@/layouts/MyPageLayout";
